use js object instead of data-attributes

+++ static/js/script.js
@@ -1,27 +1,114 @@
-document.addEventListener("DOMContentLoaded", () => {
+const backendUrl = 'https://shingenbeta.onrender.com';
+// const backendUrl = 'http://127.0.0.1:5000'
+
+/**
+ * Calls sake endpoint to get JSON data that gets transformed to a JS object
+ */
+async function fetchSakeData() {
+	try {
+		const response = await fetch(`${backendUrl}/sake`); // Adjust endpoint as necessary
+		if (!response.ok) {
+			throw new Error('Failed to fetch sake data');
+		}
+		const sakeData = await response.json();
+		return sakeData;
+	} catch (error) {
+		console.error('Error fetching sake data:', error);
+		return [];
+	}
+}
+
+/**
+ * Populate the learn more modal with selected sake details
+ */
+function populateLearnMoreModal(selected) {
+	document.getElementById("learnMoreTitle").textContent = selected.name;
+	document.getElementById("learnMoreImage").src = `data:image/jpeg;base64,${selected.image_base64}`;
+	document.getElementById("learnMorePrice").textContent = `Price: ${selected.properties.Price || 'N/A'}`;
+	document.getElementById("learnMoreSizes").textContent = `Sizes: ${selected.properties.Sizes || 'N/A'}`;
+	document.getElementById("learnMoreIn_stock").innerHTML = selected.properties['In stock'] === "True" ? 'In Stock: &check;' : 'In Stock: &#10060;';
+	document.getElementById("learnMoreExpected_date").textContent = selected.properties['In stock'] === "False" ? `Expected Date: ${selected.properties['Expected date']}` : '';
+
+	for (const [key, value] of Object.entries(selected.properties)) {
+		if (key === "In stock" || key === "Expected date") continue;
+		const elementId = `learnMore${key.replace(' ', '_')}`;
+		const element = document.getElementById(elementId);
+		if (element) {
+			element.textContent = `${key}: ${value}`;
+		}
+	}
+
+	document.getElementById("learnMoreDescription").innerText = selected.description;
+}
+
+// Render sake items using the fetched data
+function renderSakeItems(sakes) {
+	const gridContainer = document.querySelector('.grid-container');
+	gridContainer.innerHTML = ''; // Clear existing items
+
+	sakes.forEach(sake => {
+		const gridItem = document.createElement('div');
+		gridItem.classList.add('grid-item', 'learnMore');
+
+		// Add the data-id attribute with the sake's unique identifier
+		gridItem.setAttribute('data-id', sake._id);
+
+		gridItem.innerHTML = `
+					<img src="data:image/jpeg;base64,${sake.image_base64}" alt="${sake.name}">
+					<div class="content">
+							<h2>${sake.name}</h2>
+							<p>${sake.description}</p>
+					</div>
+					<div class="price">Price: ${sake.properties.Price || 'N/A'}</div>
+					<button class="btn btn-dark learnMore">&#9432;</button>
+			`;
+
+		// Add edit and delete buttons if the user is authenticated
+		if (currentUserIsAuthenticated) {
+			const editButton = document.createElement('button');
+			editButton.classList.add('btn', 'btn-dark', 'border', 'editButton');
+			editButton.textContent = 'Edit';
+
+			const deleteButton = document.createElement('button');
+			deleteButton.classList.add('btn', 'btn-dark', 'border', 'deleteButton');
+			deleteButton.textContent = 'Delete';
+
+			gridItem.appendChild(editButton);
+			gridItem.appendChild(deleteButton);
+		}
+
+		// If it's new, then add new button on top right
+		if (sake.new === 'on') {
+			const newIcon = document.createElement('div');
+			newIcon.classList.add('newIcon');
+			newIcon.textContent = "new"
+			gridItem.appendChild(newIcon)
+		}
+
+		gridContainer.appendChild(gridItem);
+	});
+}
+
+document.addEventListener("DOMContentLoaded", async () => {
 	const modal = document.getElementById("myModal");
 	const learnMoreModal = document.getElementById("learnMoreModal");
 	const closeButtons = document.querySelectorAll(".close");
 	const editForm = document.getElementById("editForm");
-	const learnMoreContent = document.getElementById("learnMoreContent");
-	// const backendUrl = 'https://shingenbeta.onrender.com';
-	const backendUrl = 'http://127.0.0.1:5000'
-	const gridItems = Array.from(document.querySelectorAll(".grid-item"));
-	const loadMoreButton = document.getElementById("loadMore");
-	// const itemsPerPage = 40;
-	// let currentPage = 1;
+
 	const clearFiltersButton = document.getElementById("clearFilters");
 	const mobileClearFiltersButton = document.getElementById("mobileClearFilters")
-	const burgerIcon = document.querySelector(".burger-icon");
 
 	const searchInput = document.querySelector("#searchInput");
 	const mobileSearchInput = document.querySelector("#mobileSearchInput");
+	const sortDropdown = document.querySelector("#sortDropdown");
 
 	// Get all sidebar and mobile checkboxes
 	const sidebarFilters = document.querySelectorAll('.sidebar input[type="checkbox"]');
 	const mobileFilters = document.querySelectorAll('.offcanvas-body input[type="checkbox"]');
 
-
+	const sakeData = await fetchSakeData();
+	console.log('SAKE DATA: ', sakeData)
+	renderSakeItems(sakeData)
 
 	// SIDEBAR ACCORDION
 	var acc = document.getElementsByClassName("accordion");
@@ -43,29 +130,6 @@ document.addEventListener("DOMContentLoaded", () => {
 		});
 	}
 
-	// 08 11 Might need to reimplement if I do lazy loading/caching
-	// But for now we compress images stored in mongoDB is sufficient
-	// function displayItems() {
-	// 	const start = (currentPage - 1) * itemsPerPage;
-	// 	const end = currentPage * itemsPerPage;
-	// 	gridItems.slice(0, end).forEach(item => item.style.display = "flex");
-	// 	gridItems.slice(end).forEach(item => item.style.display = "none");
-
-	// 	if (end >= gridItems.length) {
-	// 		loadMoreButton.style.display = "none";
-	// 	} else {
-	// 		loadMoreButton.style.display = "block";
-	// 	}
-	// }
-
-	// loadMoreButton.addEventListener("click", () => {
-	// 	currentPage += 1;
-	// 	displayItems();
-	// })
-
-	// // Initial display;
-	// displayItems();
-
 	closeButtons.forEach(button => {
 		button.onclick = () => {
 			button.closest(".modal").style.display = "none";
@@ -82,14 +146,40 @@ document.addEventListener("DOMContentLoaded", () => {
 		document.body.style.overflow = "";
 	}
 
-	// Delete a sake from DB
-	document.querySelectorAll('.deleteButton').forEach(button => {
-		button.addEventListener('click', event => {
+	// Use event delegation for grid items
+	const gridContainer = document.querySelector('.grid-container');
+
+	gridContainer.addEventListener('click', (event) => {
+		const gridItem = event.target.closest('.grid-item');
+		if (!gridItem) return; // If click is not on a grid-item, do nothing
+
+		const sakeId = gridItem.getAttribute('data-id');
+		const selected = sakeData.find(sake => sake._id === sakeId);
+
+		console.log(event.target)
+
+		// Handle Learn More button
+		if (event.target.classList.contains('learnMore') || event.target.closest('.content')) {
 			event.stopPropagation();
-			const confirmDelete = confirm("Are you sure you want to delete?")
+			learnMoreModal.style.display = "block";
+			document.body.style.overflow = 'hidden';
+			// Populate learn more modal with selected sake details
+			populateLearnMoreModal(selected);
+		}
+
+		// Handle Edit button
+		if (event.target.classList.contains('editButton')) {
+			event.stopPropagation();
+			modal.style.display = "block";
+			// Populate edit form with selected sake details
+			populateEditForm(selected);
+		}
+
+		// Handle Delete button
+		if (event.target.classList.contains('deleteButton')) {
+			event.stopPropagation();
+			const confirmDelete = confirm("Are you sure you want to delete?");
 			if (confirmDelete) {
-				const gridItem = event.target.closest('.grid-item');
-				const sakeId = gridItem.getAttribute('data-id');
 				fetch(`${backendUrl}/sake/${sakeId}`, {
 					method: 'DELETE'
 				})
@@ -102,82 +192,52 @@ document.addEventListener("DOMContentLoaded", () => {
 					})
 					.then(data => {
 						console.log('Delete response:', data);
-						// Optionally remove the grid item from the DOM
 						if (data.message === 'Sake deleted') {
 							gridItem.remove();
-							// Refresh the page to reflect changes in the database
-							location.reload();
 						}
 					})
 					.catch(error => {
 						console.error('There was a problem with the fetch operation:', error);
 					});
 			}
-		})
-	})
+		}
+	});
 
-	// Edit form handling, loads data from card into edit form
-	document.querySelectorAll('.editButton').forEach(button => {
-		button.addEventListener('click', function (event) {
-			event.stopPropagation();
-			modal.style.display = "block";
-			const contentContainer = this.parentNode.querySelector('.content');
-			if (!contentContainer) return;
+	// Populate the edit form with selected sake details
+	function populateEditForm(selected) {
+		console.log(selected)
+		editForm.name.value = selected.name || '';
+		editForm.region.value = selected.properties.Region || '';
+		editForm.brewery.value = selected.properties.Brewery || '';
+		editForm.sizes.value = selected.properties.Sizes || '';
+		editForm.taste.value = selected.properties.Taste || '';
+		editForm.pairing.value = selected.properties.Pairing || '';
+		editForm.style.value = selected.properties.Style || '';
+		editForm.price.value = selected.properties.Price || '';
+		editForm.alchohol.value = selected.properties.Alchohol || '';
+		editForm.riceType.value = selected.properties['Rice type'] || '';
+		editForm.polish.value = selected.properties.Polish || '';
+		editForm.fermentationStyle.value = selected.properties['Fermentation style'] || '';
+		editForm.body.value = selected.properties.Body || '';
+		editForm.description.value = selected.description || '';
+		editForm.inStock.checked = selected.properties['In stock'] === "True";
+		editForm.expectedDate.value = selected.properties['Expected date'] || '';
+		editForm.sakeId.value = selected._id;
+		editForm.new.checked = selected.new === 'on';
 
-			const title = contentContainer.querySelector('h2');
-			const properties = contentContainer.querySelectorAll('p');
+		const descriptionCharLength = editForm.description.value.length;
+		document.querySelector("#count").textContent = `${descriptionCharLength} / 230`;
 
-			if (title) editForm.name.value = title.textContent;
-
-			properties.forEach((p, index) => {
-				// Skip the last one since that's for description and it has no :
-				if (index < properties.length - 1) {
-					const text = p.textContent;
-					const [key, value] = text.split(': ');
-					const inputElement = editForm.querySelector(`[name="${key.toLowerCase()}"]`);
-					if (inputElement) {
-						if (inputElement.type === 'number') {
-							inputElement.value = parseFloat(value) || 0; // Use parseFloat to convert to number
-						}
-						// Dealing with in stock checkbox
-						else if (key.toLowerCase() === "in stock") {
-							if (value === "False") {
-								inputElement.checked = true;
-							} else {
-								inputElement.checked = false;
-							}
-						}
-						else {
-							inputElement.value = value;
-						}
-					}
-				}
-			});
-
-			// Set description seperately since it doesn't have :
-			editForm.querySelector('[name="description"]').value = properties[properties.length - 1].textContent
-
-			// Set description character count
-			const descriptionCharLength = document.querySelector("#description").value.length;
-			document.querySelector("#count").textContent = descriptionCharLength + " / 230";
-			// update Description char count
-			document.querySelector("#description").addEventListener('input', function () {
-				let charCount = this.value.length;
-				const count = this.nextElementSibling;
-
-				count.innerHTML = charCount + " / 230"
-				if (charCount > 230) {
-					this.value = this.value.substring(0, 230)
-					count.innerHTML = "230 / 230"
-				}
-			})
-
-			// Set the sakeId to be passed to PUT request
-			const gridItem = event.target.closest('.grid-item');
-			const sakeId = gridItem.getAttribute('data-id');
-			editForm.sakeId.value = sakeId;
-		})
-	})
+		editForm.description.addEventListener('input', function () {
+			let charCount = this.value.length;
+			const count = this.nextElementSibling;
+			count.innerHTML = `${charCount} / 230`;
+			if (charCount > 230) {
+				this.value = this.value.substring(0, 230);
+				count.innerHTML = "230 / 230";
+			}
+		});
+	}
 
 	// HANDLE EDIT SUBMIT FORM
 	editForm.addEventListener("submit", async (event) => {
@@ -203,7 +263,8 @@ document.addEventListener("DOMContentLoaded", () => {
 				"In stock": formData.get('in stock') || "True",
 				"Expected date": formData.get('expected date')
 			},
-			"description": formData.get('description')
+			"description": formData.get('description'),
+			"new": formData.get('new')
 		};
 
 		// Append JSON stringified data to FormData
@@ -231,63 +292,6 @@ document.addEventListener("DOMContentLoaded", () => {
 
 	const learnMoreClose = document.getElementsByClassName("close")[1];
 	learnMoreClose.onclick = () => learnMoreModal.style.display = "none";
-
-	/**
-	 * Handle clicking Learn More button for each grid-item
-	 * 
-	 * The idea here is that when user presses learn more, we get the info from
-	 * the nearest grid-item, and then we update the modals values with those.
-	 */
-	document.querySelectorAll('.learnMore').forEach(button => {
-		button.addEventListener('click', event => {
-			event.stopPropagation();
-			learnMoreModal.style.display = "block";
-
-			document.body.style.overflow = 'hidden'; // Disable all scrolling
-
-			const contentContainer = event.target.closest('.grid-item');
-
-			if (!contentContainer) return;
-
-			const title = contentContainer.querySelector('h2').textContent;
-			const properties = contentContainer.querySelectorAll('p');
-			const image = contentContainer.querySelector('img').src;
-
-			if (title) document.getElementById("learnMoreTitle").textContent = title;
-			document.getElementById("learnMoreImage").src = image;
-
-			properties.forEach(p => {
-				const [key, value] = p.textContent.split(': ');
-				const elementId = `learnMore${key.replace(' ', '_')}`;
-				// console.log(elementId, value);
-
-				const element = document.getElementById(elementId);
-				if (element) {
-					if (elementId === "learnMoreIn_stock") {
-						if (value.trim() === "True") {
-							element.innerHTML = `${key}: &check;`
-						} else {
-							element.innerHTML = `${key}: &#10060;`
-						}
-					}
-					else if (elementId === "learnMoreExpected_date") {
-						if (value.trim() === "") {
-							element.style.display = "none"
-						} else {
-							element.style.display = "block";
-							element.textContent = `${key}: ${value.trim()}`;
-						}
-					}
-					else {
-						element.textContent = `${key}: ${value.trim()}`;
-					}
-				}
-			});
-
-			// set description seperately since it doesn't have :
-			document.getElementById("learnMoreDescription").innerText = properties[properties.length - 1].textContent;
-		})
-	})
 
 	// Add filter functionality to each input checkbox
 	document.querySelectorAll('.filterCheckbox').forEach(checkbox => {
@@ -331,53 +335,104 @@ document.addEventListener("DOMContentLoaded", () => {
 	syncSearchInputs(mobileSearchInput, searchInput);
 
 	/**
-	 * Filter categories
-	 * 
-	 * Whenever new filter categories get added, add them into here
-	 */
-	const categories = ["region", "brewery", "alchohol", "taste", "style", "riceType", "fermentationStyle", "polish", "body"]
+ * Filter categories
+ * 
+ * Whenever new filter categories get added, include them here
+ */
+	const categories = ["Region", "Brewery", "Alchohol", "Taste", "Style", "Rice type", "Fermentation style", "Polish", "Body"];
 
 	function filterSakes() {
-		// Get the search bar value, as well as any checkedbox values selected
+		// Get the search bar value and selected checkbox values
 		const searchValue = searchInput.value.toLowerCase();
 
+		// Create an object that stores the selected values for each category
 		const selectedValues = categories.reduce((acc, category) => {
-			acc[category] = Array.from(document.querySelectorAll(`input[type="checkbox"][name="${category}"]:checked`)).map(checkbox => checkbox.value)
-
-			return acc
+			acc[category] = Array.from(document.querySelectorAll(`input[type="checkbox"][name="${category.toLowerCase()}"]:checked`)).map(checkbox => checkbox.value.toLowerCase());
+			return acc;
 		}, {});
 
-		console.log("selected values: ", selectedValues)
+		console.log("Selected values:", selectedValues);
 
-		gridItems.forEach(item => {
+		// Filter sakeData based on the search input and selected filters
+		const filteredSakes = sakeData.filter(sake => {
 			let matchesAllCategories = true;
 
+			// Check each category against selected filter values
 			categories.forEach(category => {
-				const itemValue = item.getAttribute(`data-${category}`)
-				const selectedCategoryValues = selectedValues[category]
+				const itemValue = sake.properties[category];
+				const selectedCategoryValues = selectedValues[category];
 
-				// Check if there are no selected values or if the item's attribute value matches any of the selected values
+				// Check if the item value matches any of the selected values
 				if (selectedCategoryValues.length > 0) {
-					if (category === "alchohol") {
-						// Special handling for 'alchohol' to compare numerical values
-						matchesAllCategories = matchesAllCategories && selectedCategoryValues.every(val => Number(itemValue) > val);
+					if (category === "Alchohol") {
+						// Special handling for 'Alchohol' to compare numerical values
+						matchesAllCategories = matchesAllCategories && selectedCategoryValues.some(val => Number(itemValue) > val);
 					} else {
-						matchesAllCategories = matchesAllCategories && selectedCategoryValues.includes(itemValue);
+						matchesAllCategories = matchesAllCategories && selectedCategoryValues.includes(itemValue.toLowerCase());
 					}
 				}
-			})
+			});
 
 			// Check if the sake name matches the search input
-			const sakeName = item.querySelector("h2").textContent.toLowerCase();
+			const sakeName = sake.name.toLowerCase();
 			const searchMatch = sakeName.includes(searchValue);
 
-			if (matchesAllCategories && searchMatch) {
-				item.style.display = 'flex';
-			} else {
-				item.style.display = 'none';
-			}
+			return matchesAllCategories && searchMatch;
 		});
+
+		// Render the filtered sake items
+		renderSakeItems(filteredSakes);
 	}
+
+	sortDropdown.addEventListener("change", () => {
+		const sortOption = sortDropdown.value;
+		console.log(sortOption);
+		sortSakes(sortOption);
+	});
+
+	/**
+	 * Sorts the sakes based on the dropdown input value.
+	 */
+	function sortSakes(option) {
+		// Use the sakeData array directly to sort and render
+		let sortedSakeData = [...sakeData]; // Copy sakeData to avoid mutating the original array
+
+		switch (option) {
+			case 'new':
+				console.log('sorting by new')
+				// Sort by the 'new' property first, then by any other criteria if needed
+				sortedSakeData.sort((a, b) => {
+					// If 'new' property exists and is set to 'on', place it first
+					const aIsNew = a.new === 'on' ? 1 : 0;
+					const bIsNew = b.new === 'on' ? 1 : 0;
+
+					// Sort 'new' items first; if equal, keep original order
+					return bIsNew - aIsNew;
+				});
+				break;
+			case 'aToZ':
+				sortedSakeData.sort((a, b) => a.name.localeCompare(b.name));
+				break;
+			case 'zToA':
+				sortedSakeData.sort((a, b) => b.name.localeCompare(a.name));
+				break;
+			case 'hToL':
+				sortedSakeData.sort((a, b) => parseFloat(b.properties.Price) - parseFloat(a.properties.Price));
+				break;
+			case 'lToH':
+				sortedSakeData.sort((a, b) => parseFloat(a.properties.Price) - parseFloat(b.properties.Price));
+				break;
+			case 'relevance': // Reset to the initial order
+				sortedSakeData = [...sakeData]; // Revert to the original order
+				break;
+			default:
+				break;
+		}
+
+		// Re-render the sorted items
+		renderSakeItems(sortedSakeData);
+	}
+
 
 	/**
 	 * Resets all the filters by setting search input to empty string and setting all
